Decimal price

+++ database/migrations/2024_05_16_105801_create_books_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->string('name');
             $table->text('description');
-            $table->unsignedInteger('price');
+            $table->decimal('price', 10, 2);
             $table->string('author');
             $table->timestamps();
         });
